+++ R/prediction.R
@@ -62,6 +62,8 @@ gcap.runPrediction <- function(data,
 
   lg$info("selecting necessary features from input data")
   if (is.null(data$pLOH)) data$pLOH <- NA
+  if (is.null(data$age)) data$age <- NA
+  if (is.null(data$gender)) data$gender <- NA
   data <- tryCatch(
     as.matrix(data[, model$feature_names, with = FALSE]),
     error = function(e) {