+++ R/workflow.R
@@ -178,25 +178,26 @@ gcap.workflow <- function(tumourseqfile, normalseqfile,
     model = model
   )
 
-  save_file <- file.path(outdir, paste0(result_file_prefix, "_prediction_result.csv"))
-  lg$info("Saving result to {save_file}")
-  data.table::fwrite(model_input, file = save_file)
+  save_file <- file.path(outdir, paste0(result_file_prefix, "_prediction_result.rds"))
+  lg$info("Saving raw prediction result to {save_file}")
+  saveRDS(model_input, file = save_file)
 
   lg$info("====================================")
   lg$info("Step 4: Run scoring and summarizing")
   lg$info("====================================")
-  out <- gcap.runScoring(model_input)
+  out <- gcap.runScoring(model_input, genome_build)
 
-  save_file <- file.path(outdir, paste0(result_file_prefix, "_by_case.csv"))
-  lg$info("Saving result to {save_file}")
-  data.table::fwrite(out, file = save_file)
+  save_file <- file.path(outdir, paste0(result_file_prefix, "_by_gene.csv"))
+  lg$info("Saving gene result to {save_file}")
+  data.table::fwrite(out$gene, file = save_file)
+
+  save_file <- file.path(outdir, paste0(result_file_prefix, "_by_sample.csv"))
+  lg$info("Saving sample result to {save_file}")
+  data.table::fwrite(out$sample, file = save_file)
 
   lg$info("=======================================")
   lg$info(" Done! Thanks for using GCAP workflow")
   lg$info("=======================================")
 
-  invisible(list(
-    by_gene = model_input,
-    by_case = out
-  ))
+  invisible(out)
 }