+++ NEWS.md
@@ -1,6 +1,7 @@
 # gcap 0.4.1
 
-- Automatically appended logs to specific directoy with `rappdirs::app_dir("gcap", "ShixiangWang")`.
+- Filled `NA`s to input when age and gender are not available.
+- Automatically appended logs to specific directory with `rappdirs::app_dir("gcap", "ShixiangWang")`.
 Users can obtain log path and cat log info with `gcap:::get_log_file()` and `gcap:::cat_log_file()`
 for debugging. (#14)
 - Supported XGB54 model in workflows. (#13)